15HW add put and delete for post

+++ src/bloggers/repository/bloggers.repository.ts
@@ -1,18 +1,15 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
 import {
   Blog,
   BlogsDocument,
   BlogsModelType,
   BlogsViewModel,
 } from '../schemas/blogs.schema';
-import { CreateBlogDto } from '../dto/input-blog.dto';
-import * as mongoose from 'mongoose';
-import ObjectId = mongoose.Types.ObjectId;
+import { CreateBlogDto } from '../dto/input-bloggers.dto';
 
 @Injectable()
-export class BlogsRepository {
+export class BloggersRepository {
   constructor(
     @InjectModel(Blog.name)
     private readonly blogModel: BlogsModelType,
@@ -26,7 +23,7 @@ export class BlogsRepository {
     userId: string,
   ): Promise<BlogsDocument> {
     return this.blogModel.findOneAndUpdate(
-      { id: blogId, userId: userId },
+      { id: blogId, 'blogOwnerInfo.userId': userId },
       {
         name: blog.name,
         description: blog.description,
@@ -35,6 +32,6 @@ export class BlogsRepository {
     );
   }
   async deleteBlogById(id: string, userId: string) {
-    return this.blogModel.deleteOne({ id, userId: userId });
+    return this.blogModel.deleteOne({ id, 'blogOwnerInfo.userId': userId });
   }
 }
